+++ derive/src/context/num/traits/cast.rs
@@ -16,18 +16,14 @@ impl Context {
             num_traits,
             modtype,
             struct_ident,
-            generics,
             field_ident,
             ..
         } = self;
-        let generics = self.with_features(
-            &[
-                parse_quote!(Subtraction),
-                parse_quote!(Multiplication),
-                parse_quote!(Division),
-            ],
-            &generics,
-        );
+        let generics = self.with_features(&[
+            parse_quote!(PartialSubtraction),
+            parse_quote!(PartialMultiplication),
+            parse_quote!(PartialDivision),
+        ]);
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
         let struct_expr = self.struct_expr(true, None);