+++ derive/src/context/num/traits/sign.rs
@@ -12,18 +12,14 @@ impl Context {
         let Self {
             num_traits,
             struct_ident,
-            generics,
             ..
         } = self;
-        let generics = self.with_features(
-            &[
-                parse_quote!(Addition),
-                parse_quote!(Subtraction),
-                parse_quote!(Multiplication),
-                parse_quote!(Division),
-            ],
-            &generics,
-        );
+        let generics = self.with_features(&[
+            parse_quote!(PartialAddition),
+            parse_quote!(PartialSubtraction),
+            parse_quote!(PartialMultiplication),
+            parse_quote!(PartialDivision),
+        ]);
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
         quote! {