+++ derive/src/context/std/ops.rs
@@ -8,12 +8,11 @@ impl Context {
         let Context {
             std,
             struct_ident,
-            generics,
             field_ident,
             field_ty,
             ..
         } = self;
-        let generics = self.with_features(&[parse_quote!(Deref)], &generics);
+        let generics = self.with_features(&[parse_quote!(Deref)]);
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
         quote! {
@@ -35,11 +34,10 @@ impl Context {
             modulus,
             std,
             struct_ident,
-            generics,
             field_ident,
             ..
         } = self;
-        let generics = self.with_features(&[parse_quote!(Subtraction)], &generics);
+        let generics = self.with_features(&[parse_quote!(PartialSubtraction)]);
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
         let (struct_update, struct_update_deref) = self.struct_update(
@@ -73,7 +71,11 @@ impl Context {
     }
 
     pub(crate) fn derive_add(&self) -> proc_macro2::TokenStream {
-        self.derive_bin(parse_quote!(Add), parse_quote!(add), parse_quote!(Addition))
+        self.derive_bin(
+            parse_quote!(Add),
+            parse_quote!(add),
+            parse_quote!(PartialAddition),
+        )
     }
 
     pub(crate) fn derive_add_assign(&self) -> proc_macro2::TokenStream {
@@ -81,7 +83,7 @@ impl Context {
             parse_quote!(AddAssign),
             parse_quote!(add_assign),
             parse_quote!(add),
-            parse_quote!(Addition),
+            parse_quote!(PartialAddition),
         )
     }
 
@@ -89,7 +91,7 @@ impl Context {
         self.derive_bin(
             parse_quote!(Sub),
             parse_quote!(sub),
-            parse_quote!(Subtraction),
+            parse_quote!(PartialSubtraction),
         )
     }
 
@@ -98,7 +100,7 @@ impl Context {
             parse_quote!(SubAssign),
             parse_quote!(sub_assign),
             parse_quote!(sub),
-            parse_quote!(Subtraction),
+            parse_quote!(PartialSubtraction),
         )
     }
 
@@ -106,7 +108,7 @@ impl Context {
         self.derive_bin(
             parse_quote!(Mul),
             parse_quote!(mul),
-            parse_quote!(Multiplication),
+            parse_quote!(PartialMultiplication),
         )
     }
 
@@ -115,12 +117,16 @@ impl Context {
             parse_quote!(MulAssign),
             parse_quote!(mul_assign),
             parse_quote!(mul),
-            parse_quote!(Multiplication),
+            parse_quote!(PartialMultiplication),
         )
     }
 
     pub(crate) fn derive_div(&self) -> proc_macro2::TokenStream {
-        self.derive_bin(parse_quote!(Div), parse_quote!(div), parse_quote!(Division))
+        self.derive_bin(
+            parse_quote!(Div),
+            parse_quote!(div),
+            parse_quote!(PartialDivision),
+        )
     }
 
     pub(crate) fn derive_div_assign(&self) -> proc_macro2::TokenStream {
@@ -128,12 +134,16 @@ impl Context {
             parse_quote!(DivAssign),
             parse_quote!(div_assign),
             parse_quote!(div),
-            parse_quote!(Division),
+            parse_quote!(PartialDivision),
         )
     }
 
     pub(crate) fn derive_rem(&self) -> proc_macro2::TokenStream {
-        self.derive_bin(parse_quote!(Rem), parse_quote!(rem), parse_quote!(Division))
+        self.derive_bin(
+            parse_quote!(Rem),
+            parse_quote!(rem),
+            parse_quote!(PartialDivision),
+        )
     }
 
     pub(crate) fn derive_rem_assign(&self) -> proc_macro2::TokenStream {
@@ -141,7 +151,7 @@ impl Context {
             parse_quote!(RemAssign),
             parse_quote!(rem_assign),
             parse_quote!(rem),
-            parse_quote!(Division),
+            parse_quote!(PartialDivision),
         )
     }
 
@@ -155,11 +165,10 @@ impl Context {
             modulus,
             std,
             struct_ident,
-            generics,
             field_ident,
             ..
         } = self;
-        let generics = self.with_features(&[feature], &generics);
+        let generics = self.with_features(&[feature]);
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
         let (struct_update, struct_update_deref) = self.struct_update(
@@ -237,11 +246,10 @@ impl Context {
             modulus,
             std,
             struct_ident,
-            generics,
             field_ident,
             ..
         } = &self;
-        let generics = self.with_features(&[feature], &generics);
+        let generics = self.with_features(&[feature]);
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
         let (_, update) = self.struct_update(